+++ src/main/java/racingcar/validate/CheckCarName.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validate;
 
 import java.util.Collections;
 import java.util.List;
