+++ src/main/java/racingcar/validate/CheckPlayTime.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validate;
 
 import racingcar.constant.ConstNumber;
