+++ src/main/java/racingcar/validate/Validation.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validate;
 
 import java.util.Arrays;
 import java.util.List;
